switching responsible component
