--- frontend/src/constants/styles/device.ts
@@ -1,11 +1,0 @@
-const breakpoints = {
-  mobile: "40.0rem",
-  tablet: "80.0rem"
-};
-
-export default {
-  mobile: `screen and (max-width: ${breakpoints.mobile})`,
-  tablet: `screen and (min-width: ${breakpoints.mobile}) and (max-width: ${
-    breakpoints.tablet
-  })`
-};
